+++ NEWS.md
@@ -1,3 +1,15 @@
+## easySdcTable	0.8.0
+
+* Method "Gauss" made default. 
+  - Based on comparison of computing times, errors and the numbers of suppressed and revealed cells,
+this is a natural consequence after the release of sdcTable version 0.32.0 
+(see `Comparison of suppression methods while developing easySdcTable ver. 0.8.0.md` on the github page of easySdcTable). 
+* Internal changes caused by changes in sdcTable version 0.32.0
+  - Now `protect_linked_tables` is used insetad of `protectLinkedTables`
+  - Summary output from sdcTable is captured differently (for `info` output)
+* New tidy wrapper function, `ProtectTableData`,  which returns a single data frame. 
+
+
 ## easySdcTable	0.7.0
 
 * No changes in functionality. Only technical changes and minor improvements.